requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
wfdb>=4.1.0
sounddevice>=0.4.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0
pushbullet.py>=0.11.0
pyserial>=3.5